Upgrade to Framework Version 4.5
Here are the changes required to move the project from framework version
4.0 to framework version 4.5.

There was no problems when performing this upgrade since .NET framework
4.0 to .NET framework 4.5 is an in place upgrade. This means all
existing files are only upgraded not replaced (this is a minor release
not a major release).

As a matter of fact, there is no directory for the 4.5 framework there
is only a directory for the 4.0 version (this directory being 4.0.309
[not sure where it's located sorry])

I have done many similar upgrades at my current job. From framework 4.0
to 4.5 and between C++ toolsets (toolset v10.0 to toolset v11.0)