Scene redefines methods without GameObjectManager
By implementing the new interface, all methods were defined using the
GameObjectManager class. However, instead of calling true the
GameObjectManager object like it used to, the methods are already
available inside the Scene class, which makes it easier to remove the
reference to the GameObjectManager object. Every reference to that
object was removed except for the events, which will be handle in the
next commit.