Scene implements the IGameObjectManager again
Due to certain mistakes, the scene was reverted to its original and all
the methods and attributes from the interface were implemented for a
second time. All the method definitions were filled at the same time in
order to make it work. Problem : 2 unit tests are failing at the moment.